Split resolvers into seperate files
