+++ .vscode/settings.json
@@ -1,4 +1,4 @@
 {
-  "editor.defaultFormatter": "esbenp.prettier-vscode",
+  "editor.defaultFormatter": "dbaeumer.vscode-eslint",
   "prettier.configPath": ".prettierrc"
 }
