+++ packages/server/package.json
@@ -18,6 +18,7 @@
     "graphql": "^16.5.0",
     "graphql-fields-projection": "^1.0.7",
     "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.21",
     "mongoose": "^6.4.0",
     "validator": "^13.7.0"
   },
@@ -29,6 +30,7 @@
     "@types/express": "^4.17.13",
     "@types/jest": "^27.4.1",
     "@types/jsonwebtoken": "^8.5.8",
+    "@types/lodash": "^4.14.182",
     "@types/node": "^18.0.0",
     "@types/validator": "^13.7.3",
     "jest": "^28.1.1",
